Unified renderer
