+++ sites/svelte.dev/src/routes/docs/[slug]/+page.svelte
@@ -1,8 +1,7 @@
 <script>
 	import { page } from '$app/stores';
-	import * as hovers from '$lib/utils/hovers';
 	import { Icon } from '@sveltejs/site-kit/components';
-	import { DocsOnThisPage } from '@sveltejs/site-kit/docs';
+	import { DocsOnThisPage, setupDocsHovers } from '@sveltejs/site-kit/docs';
 
 	export let data;
 
@@ -11,7 +10,7 @@
 	$: prev = pages[index - 1];
 	$: next = pages[index + 1];
 
-	hovers.setup();
+	setupDocsHovers();
 </script>
 
 <svelte:head>
